move mock classes to the module from tests/
