+++ src/switch_network/__init__.py
@@ -1,5 +1,5 @@
 __author__ = "Charlie Tolley"
 __version__ = "0.0.1"
 
-from . import pico_utils
 from .switch import SwitchNetwork
+from . import pico_utils, testing
